move all shared dependencies into a utils folder outside of packages export this from devtools still but for core use the utils folder directly
