+++ packages/core/integrations/index.js
@@ -5,7 +5,6 @@ const { IntegrationMapping } = require('./integration-mapping');
 const { IntegrationFactory, IntegrationHelper } = require('./integration-factory');
 const { createIntegrationRouter, checkRequiredParams } = require('./integration-router');
 const { createFriggBackend } = require('./create-frigg-backend');
-const { createMockApiObject, createMockIntegration } = require('./mock-integration');
 
 module.exports = {
     IntegrationBase,
@@ -17,6 +16,4 @@ module.exports = {
     createIntegrationRouter,
     checkRequiredParams,
     createFriggBackend,
-    createMockIntegration,
-    createMockApiObject
 };
