+++ packages/core/integrations/jest-teardown.js
@@ -1,2 +1,2 @@
-const { globalTeardown } = require('@friggframework/devtools');
+const { globalTeardown } = require('../../../utils/test-environment');
 module.exports = globalTeardown;
