+++ packages/core/module-plugin/index.js
@@ -9,8 +9,6 @@ const { Requester } = require('./requester/requester');
 const { ModuleConstants } = require('./ModuleConstants');
 const { ModuleFactory } = require('./module-factory');
 const { Auther } = require('./auther');
-const { testAutherDefinition } = require('./auther-definition-tester');
-const { testDefinitionRequiredAuthMethods } = require('./auther-definition-method-tester');
 
 module.exports = {
     Credential,
@@ -23,7 +21,5 @@ module.exports = {
     Requester,
     ModuleConstants,
     ModuleFactory,
-    Auther,
-    testAutherDefinition,
-    testDefinitionRequiredAuthMethods
+    Auther
 };
