+++ packages/core/module-plugin/jest-setup.js
@@ -1,3 +1,3 @@
 require('dotenv').config();
-const { globalSetup } = require('@friggframework/devtools');
+const { globalSetup } = require('../../../utils/test-environment');
 module.exports = globalSetup;
