+++ packages/core/package.json
@@ -15,7 +15,6 @@
     "node-fetch": "^2.6.7"
   },
   "devDependencies": {
-    "@friggframework/devtools": "^1.1.0",
     "@types/lodash": "^4.14.191",
     "@typescript-eslint/eslint-plugin": "^5.55.0",
     "chai": "^4.3.6",
