+++ packages/devtools/index.js
@@ -1,5 +1,5 @@
-const eslintConfig = require('./eslint-config')
-const prettierConfig = require('./prettier-config')
+const eslintConfig = require('../../../utils/eslint-config')
+const prettierConfig = require('../../../utils/prettier-config')
 const testEnvironment = require('./test-environment/index');
 
 module.exports = {
