+++ packages/devtools/test-environment/auther-definition-method-tester.js
@@ -1,4 +1,4 @@
-const {flushDebugLog} = require('../logs');
+const {flushDebugLog} = require('@friggframework/core');
 
 async function testDefinitionRequiredAuthMethods(api, definition, authCallbackParams, tokenResponse, userId) {
