+++ packages/devtools/test-environment/index.js
@@ -1,21 +1,25 @@
-const { TestMongo } = require('./mongodb');
+const Authenticator = require('@friggframework/devtools/test/Authenticator')
+const { createMockIntegration, createMockApiObject } = require('mock-integration');
+const { testAutherDefinition } = require('./auther-definition-tester');
+const { testDefinitionRequiredAuthMethods } = require('./auther-definition-method-tester');
+const {  } = require('./../../utils/test-environment');
 const {
-    overrideEnvironment,
-    restoreEnvironment,
-} = require('./override-environment');
-const globalTeardown = require('./jest-global-teardown');
-const globalSetup = require('./jest-global-setup');
-const Authenticator = require('./Authenticator')
-
-module.exports = {
     TestMongo,
     overrideEnvironment,
     restoreEnvironment,
     globalTeardown,
     globalSetup,
+} = require('./../../../utils/test-environment');
+
+module.exports = {
     createMockIntegration,
     createMockApiObject,
     testDefinitionRequiredAuthMethods,
     testAutherDefinition,
-    Authenticator
+    Authenticator,
+    TestMongo,
+    overrideEnvironment,
+    restoreEnvironment,
+    globalTeardown,
+    globalSetup,
 };
