+++ utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@friggframework/devtools",
+  "name": "@friggframework/utils",
   "prettier": "@friggframework/prettier-config",
   "version": "1.1.0",
   "dependencies": {
@@ -14,7 +14,6 @@
     "mongodb-memory-server": "^8.9.0"
   },
   "devDependencies": {
-    "@friggframework/core": "^1.1.0",
     "jest": "^28.1.3",
     "prettier": "^2.7.1"
   },
